requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.35.0
httpx==0.27.0
email-validator==2.1.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
SQLAlchemy==2.0.32
alembic==1.13.2
google-api-python-client==2.139.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytesseract==0.3.10
anthropic==0.34.2
pdf2image==1.17.0
apscheduler==3.10.4
Pillow==10.0.1
